refactor: workers by system
