+++ packages/backend/src/workers/WindowsWorker.ts
@@ -1,0 +1,27 @@
+/**********************************************************************
+ * Copyright (C) 2024 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+import { env } from '@podman-desktop/api';
+import type { IWorker } from './IWorker';
+
+export abstract class WindowsWorker<T, R> implements IWorker<T, R> {
+  enabled(): boolean {
+    return env.isWindows;
+  }
+
+  abstract perform(content: T): Promise<R>;
+}
